RISDEV-12577 Improve tests
* Refactor tests that don't use the standard data into different classes to make tests not order dependent.
